refactor: rename directories & enums
